+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0
+
+* Add AsyncCascadeView widget
+* Change selected item background color to Theme.of(context).colorScheme.primary
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* initial release.